Did some major coding.
Added a modelfactory which returns the current model to be used, made "LargeHerbivore" an abstract class and the three herbivores extending that class. Made a NaturePreserve class which has a method getState which return the state of the NaturePreserve and all of the species in it after giving it a certain time. Some major bugs still need to be fixed